DoctrineDataSource - added support for Doctrine 3+.
