+++ Clockwork/DataSource/DoctrineDataSource.php
@@ -5,8 +5,8 @@ use Doctrine\ORM\EntityManager;
 // Data source for Doctrine, provides database queries
 class DoctrineDataSource extends DBALDataSource
 {
-	public function __construct(EntityManager $enm)
+	public function __construct(?EntityManager $enm = null)
 	{
-		parent::__construct($enm->getConnection());
+		parent::__construct($enm ? $enm->getConnection() : null);
 	}
 }
